Add new feature of pre-defined shortcut to specify SID and client directly
An user sometimes requires to open new session that is not defined in ini file.
e.g. The system and/or client is used rarely.

However, in that case, it annoys for picking up the SID from Logon-pad.
So this commit adds a feature to open new session by simply putting a SID and a client as a shortcut.

Caveat:
This feature never checked the existence/validity of a SID and a client.
It only checks the length and the character type.